% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALDreg.R
\name{ALDreg}
\alias{ALDreg}
\title{\code{ALDreg} fits density regression for longitudinal data assuming asymmetric Laplace distribution}
\usage{
ALDreg(
  formFixed,
  formRandom,
  formGroup,
  formFixedScale,
  formRandomScale,
  formFixedSkewness = NULL,
  formRandomSkewness = NULL,
  Skewness_link = "logit",
  data,
  n.chains = 3,
  n.iter = 6000,
  n.burnin = 1000,
  n.thin = 1,
  n.adapt = NULL,
  object_lslqmm = NULL,
  precision = 0.1,
  save_jagsUI = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{formFixed}{formula for fixed part of location regression with response variable}

\item{formRandom}{formula for random part of location regression without left formula part}

\item{formGroup}{formula specifying the cluster variable (e.g. = ~ subject)}

\item{formFixedScale}{formula for fixed part of scale regression without left formula part (log link)}

\item{formRandomScale}{formula for random part of scale regression without left formula part  (log link)}

\item{formFixedSkewness}{formula for fixed part of Skewness regression without left formula part (logit link)}

\item{formRandomSkewness}{formula for random part of Skewness regression without left formula part (logit link)}

\item{Skewness_link}{Considered link between the linear predictor and the skewness parameter. By default "logit", but considering only group-specific constant, the beta distribution can use as prior by specifying "beta".}

\item{data}{long format dataset of observed variables}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 10000}

\item{n.burnin}{integer specifying how many of \code{n.iter} to discard as burn-in ; default is 5000}

\item{n.thin}{integer specifying the thinning of the chains; default is 1}

\item{n.adapt}{integer specifying the number of iterations to use for adaptation; default is \code{NULL}}

\item{object_lslqmm}{Object returns by 'lslqmm' function of BeQuT package to initialize parameter chains, default is \code{NULL}}

\item{precision}{precision (e.g. inverse of variance) of prior distributions, default is 0.1}

\item{save_jagsUI}{If \code{TRUE} (by default), the output of \code{jagsUI} package is returned by the function. Warning, if \code{TRUE}, the output can be large.}

\item{parallel}{see \code{jagsUI::jags()} function}
}
\value{
A \code{BALDreg} object is a list with the following elements:
\describe{
\item{\code{mean}}{list of posterior mean for each parameter}
\item{\code{median}}{list of posterior median for each parameter}
\item{\code{modes}}{list of posterior mode for each parameter}
\item{\code{StErr}}{list of standard error for each parameter}
\item{\code{StDev}}{list of standard deviation for each parameter}
\item{\code{Rhat}}{list of Rhat convergence criteria for each parameter}
\item{\code{ICs}}{list of the credibility interval at 0.95 for each parameters excepted for covariance parameters in covariance matrix of random effects. Otherwise, use save_jagsUI=TRUE to have the associated quantiles.}
\item{\code{data}}{data included in argument}
\item{\code{sims.list}}{list of the MCMC chains of the parameters and random effects}
\item{\code{control}}{list of arguments giving details about the estimation}
\item{\code{random_effect_b}}{list including posterior mean, posterior standard deviation and Rhat criteria of subject-specific location random effects}
\item{\code{random_effect_u}}{list including posterior mean, posterior standard deviation and Rhat criteria of subject-specific scale random effects}
\item{\code{random_effect_a}}{list including posterior mean, posterior standard deviation and Rhat criteria of subject-specific skewness random effects}
\item{\code{random_effect_tau}}{only if no covariates are considered in skewness part, list including posterior mean, posterior standard deviation and Rhat criteria of subject-specific skewness random effect}
\item{\code{out_jagsUI}}{only if \code{save_jagsUI=TRUE} in argument: list including posterior mean, median, quantiles (2.5\%, 25\%, 50\%, 75\%, 97.5\%), standart deviation for each parameter and each random effect.
Moreover, this list also returns the MCMC draws, the Gelman and Rubin diagnostics (see output of jagsUI objects)}
}
}
\description{
Function using 'JAGS' software to estimate the individual response distribution assuming asymmetric Laplace
distribution.
}
\examples{

\donttest{
#---- Load data
data(threeC)

#---- data management
threeC$age.visit65 <- (threeC$age.visit-65)/10
threeC$age.final65 <- (threeC$age.final-65)/10
threeC$age0_65 <- (threeC$age0-65)/10
threeC$age.last65 <- (threeC$age.last-65)/10
threeC$age.first65 <- (threeC$age.first-65)/10
threeC$SBP <- threeC$SBP/10

#---- Fit ALD regression model with subject-specific random effect for skewness part
model_beta <- ALDreg(formFixed = SBP ~ age.visit65,
                        formRandom = ~ age.visit65,
                        formFixedScale = ~ age.visit65,
                        formRandomScale = ~ age.visit65,
                        formGroup = ~ ID,
                        Skewness_link = "beta",
                        data = threeC,
                        n.iter = 3000,
                        n.burnin = 1000)

#---- Fit ALD regression model with subject-specific random effect for skewness part
model_intercept <- ALDreg(formFixed = SBP ~ age.visit65,
                        formRandom = ~ age.visit65,
                        formFixedScale = ~ age.visit65,
                        formRandomScale = ~ age.visit65,
                        formGroup = ~ ID,
                        Skewness_link = "logit",
                        data = threeC,
                        n.iter = 3000,
                        n.burnin = 1000)

#---- Get the posterior means
model_intercept$mean

#---- Get the posterior prediction of subject-specific skewness part
model_intercept$out$random_effect_tau$postMeans

#---- Visualize the trace for beta parameters
jagsUI::traceplot(model_intercept$out_jagsUI, parameters = "beta")

#---- Summary of output
summary(model_intercept)

#---- Fit full ALD regression model with subject-specific random effect defining all parts
model_full <- ALDreg(formFixed = SBP ~ age.visit65,
                     formRandom = ~ age.visit65,
                     formFixedScale = ~ age.visit65,
                     formRandomScale = ~ age.visit65,
                     formFixedSkewness = ~ age.visit65,
                     formRandomSkewness = ~ age.visit65,
                     formGroup = ~ ID,
                     data = threeC,
                     n.iter = 3000,
                     n.burnin = 1000)
}

}
\references{
Courcoul L, Tzourio C, Woodward M, Barbieri A, Jacqmin-Gadda H (2024).
\emph{	A location-scale joint model for studying the link between the time-dependent subject-specific variability of blood pressure and competing events}.
arXiv, arXiv:2306.16785. doi: https://doi.org/10.48550/arXiv.2306.16785.
}
\author{
Antoine Barbieri
}
