% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeQut-package.R
\docType{package}
\name{BeQut-package}
\alias{BeQut}
\alias{BeQut-package}
\title{BeQut: Bayesian Estimation for Quantile Regression Mixed Models}
\description{
Using a Bayesian estimation procedure, this package fits linear quantile regression models such as linear quantile model, linear quantile mixed models, quantile regression joint model for time-to-event and longitudinal data. The estimation procedure is based on asymmetric Laplace distribution and the JAGS software is used to get posterior sample.
}
\author{
\strong{Maintainer}: Antoine Barbieri \email{antoine.barbieri@u-bordeaux.fr}

Authors:
\itemize{
  \item Hélène Jacqmin-Gadda \email{helene.jacqmin-gadda@u-bordeaux.fr}
}

}
\keyword{internal}
