% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lslqmm.R
\name{lslqmm}
\alias{lslqmm}
\title{\code{lslqmm} fits location-scale linear quantile mixed model}
\usage{
lslqmm(
  formFixed,
  formRandom,
  formGroup,
  formFixedScale,
  formRandomScale,
  data,
  tau,
  n.chains = 3,
  n.iter = 10000,
  n.burnin = 5000,
  n.thin = 1,
  n.adapt = NULL,
  precision = 10,
  save_jagsUI = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{formFixed}{formula for fixed part of longitudinal submodel with response variable}

\item{formRandom}{formula for random part of longitudinal submodel without left formula part}

\item{formGroup}{formula specifying the cluster variable (e.g. = ~ subject)}

\item{formFixedScale}{formula for fixed part of scale regression part without left formula part}

\item{formRandomScale}{formula for random part of scale regression part without left formula part}

\item{data}{long format dataset of observed variables}

\item{tau}{the quantile(s) to be estimated. This must be a number between 0 and 1, otherwise the execution is stopped. If more than one quantile is specified, rounding off to the 4th decimal must give non–duplicated values of \code{tau}, otherwise the execution is stopped.}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 10000}

\item{n.burnin}{integer specifying how many of \code{n.iter} to discard as burn-in ; default is 5000}

\item{n.thin}{integer specifying the thinning of the chains; default is 1}

\item{n.adapt}{integer specifying the number of iterations to use for adaptation; default is \code{NULL}}

\item{precision}{variance by default for vague prior distribution}

\item{save_jagsUI}{If \code{TRUE} (by default), the output of \code{jagsUI} package is returned by the function. Warning, if \code{TRUE}, the output can be large.}

\item{parallel}{see \code{jagsUI::jags()} function}
}
\value{
A \code{Blqmm} object is a list with the following elements:
\describe{
\item{\code{mean}}{list of posterior mean for each parameter}
\item{\code{median}}{list of posterior median for each parameter}
\item{\code{modes}}{list of posterior mode for each parameter}
\item{\code{StErr}}{list of standard error for each parameter}
\item{\code{StDev}}{list of standard deviation for each parameter}
\item{\code{ICs}}{list of the credibility interval at 0.95 for each parameters excepted for covariance parameters in covariance matrix of random effects. Otherwise, use save_jagsUI=TRUE to have the associated quantiles.}
\item{\code{data}}{data included in argument}
\item{\code{sims.list}}{list of the MCMC chains of the parameters and random effects}
\item{\code{control}}{list of arguments giving details about the estimation}
\item{\code{random_effect}}{list for each quantile including both posterior mean and posterior standard deviation of subject-specific random effects}
\item{\code{out_jagsUI}}{only if \code{save_jagsUI=TRUE} in argument: list including posterior mean, median, quantiles (2.5\%, 25\%, 50\%, 75\%, 97.5\%), standart deviation for each parameter and each random effect.
Moreover, this list also returns the MCMC draws, the Gelman and Rubin diagnostics (see output of jagsUI objects)}
}
}
\description{
Function using 'JAGS' software to estimate the linear quantile mixed model assuming asymmetric Laplace
distribution for residual error.
}
\examples{

\donttest{
#---- Use dataLong dataset
data(pbc2, package = "JMbayes")

#---- Fit regression model for the median
lslqmm_median <- lslqmm(formFixed = log(serBilir) ~ year,
                        formRandom = ~ year,
                        formFixedScale = ~ year,
                        formRandomScale = ~ year,
                        formGroup = ~ id,
                        data = pbc2,
                        tau = 0.5,
                        n.iter = 10000,
                        n.burnin = 1000)

#---- Get the posterior means
lslqmm_median$mean

#---- Visualize the trace for beta parameters
jagsUI::traceplot(lslqmm_median$out_jagsUI, parameters = "beta")

#---- Summary of output
summary(lslqmm_median)
}

}
\references{
Courcoul L, Tzourio C, Woodward M, Barbieri A, Jacqmin-Gadda H (2024).
\emph{	A location-scale joint model for studying the link between the time-dependent subject-specific variability of blood pressure and competing events}.
arXiv, arXiv:2306.16785. doi: https://doi.org/10.48550/arXiv.2306.16785.

Marco Geraci and Matteo Bottai (2014).
\emph{Linear quantile mixed models}.
Statistics and Computing, 24(3):461-479. doi: 10.1007/s11222-013-9381-9.
}
\author{
Antoine Barbieri
}
