% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scALDreg.R
\name{scALDreg}
\alias{scALDreg}
\title{\code{scALDreg} fits density regression for longitudinal data assuming asymmetric Laplace distribution}
\usage{
scALDreg(
  formFixedScale,
  formRandomScale,
  Location = 0,
  formGroup,
  tau,
  lambda = 0,
  data,
  n.chains = 3,
  n.iter = 6000,
  n.burnin = 1000,
  n.thin = 1,
  n.adapt = NULL,
  precision = 0.1,
  save_jagsUI = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{formFixedScale}{formula for fixed part of scale regression with left formula part is response variable (log link)}

\item{formRandomScale}{formula for random part of scale regression without left formula part  (log link)}

\item{Location}{numeric value, must be'0' or '1' to fix Location part to 0 or to fit an intercept, respectively."}

\item{formGroup}{formula specifying the cluster variable (e.g. = ~ subject)}

\item{tau}{skewness (vector) parameter. By default, this is a scalar between 0 and 1, but it can be a n-length vector specifying the individual skewness.}

\item{lambda}{numeric regarding the penanlisation on fixed scale part. By default, 0.}

\item{data}{long format dataset of observed variables}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 10000}

\item{n.burnin}{integer specifying how many of \code{n.iter} to discard as burn-in ; default is 5000}

\item{n.thin}{integer specifying the thinning of the chains; default is 1}

\item{n.adapt}{integer specifying the number of iterations to use for adaptation; default is \code{NULL}}

\item{precision}{precision (e.g. inverse of variance) of prior distributions, default is 0.1}

\item{save_jagsUI}{If \code{TRUE} (by default), the output of \code{jagsUI} package is returned by the function. Warning, if \code{TRUE}, the output can be large.}

\item{parallel}{see \code{jagsUI::jags()} function}
}
\value{
A \code{BpscALDreg} object is a list with the following elements:
\describe{
\item{\code{mean}}{list of posterior mean for each parameter}
\item{\code{median}}{list of posterior median for each parameter}
\item{\code{modes}}{list of posterior mode for each parameter}
\item{\code{StErr}}{list of standard error for each parameter}
\item{\code{StDev}}{list of standard deviation for each parameter}
\item{\code{ICs}}{list of the credibility interval at 0.95 for each parameters excepted for covariance parameters in covariance matrix of random effects. Otherwise, use save_jagsUI=TRUE to have the associated quantiles.}
\item{\code{data}}{data included in argument}
\item{\code{sims.list}}{list of the MCMC chains of the parameters and random effects}
\item{\code{control}}{list of arguments giving details about the estimation}
\item{\code{random_effect_b}}{list for each quantile including both posterior mean and posterior standard deviation of subject-specific location random effects}
\item{\code{random_effect_u}}{list for each quantile including both posterior mean and posterior standard deviation of subject-specific scale random effects}
\item{\code{random_effect_tau}}{list for each quantile including both posterior mean and posterior standard deviation of subject-specific skewness random effects}
\item{\code{out_jagsUI}}{only if \code{save_jagsUI=TRUE} in argument: list including posterior mean, median, quantiles (2.5\%, 25\%, 50\%, 75\%, 97.5\%), standart deviation for each parameter and each random effect.
Moreover, this list also returns the MCMC draws, the Gelman and Rubin diagnostics (see output of jagsUI objects)}
}
}
\description{
Function using 'JAGS' software to estimate the individual response distribution assuming asymmetric Laplace
distribution for residual error.
}
\examples{

\donttest{
#---- Use dataLong dataset
data(pbc2, package = "JMbayes")

#---- Fit regression model for the median
test <- scALDreg(formFixedScale = log(serBilir) ~ year,
                  formRandomScale = ~ year,
                  formGroup = ~ id,
                  data = pbc2,
                  tau = 0.5,
                  lambda = 0,
                  n.iter = 2000,
                  n.burnin = 1000)

#---- Get the posterior means
test$mean

#---- Visualize the trace for beta parameters
jagsUI::traceplot(test$out_jagsUI, parameters = "xi")

#---- Summary of output
summary(test)
}

}
\references{
Courcoul L, Tzourio C, Woodward M, Barbieri A, Jacqmin-Gadda H (2024).
\emph{	A location-scale joint model for studying the link between the time-dependent subject-specific variability of blood pressure and competing events}.
arXiv, arXiv:2306.16785. doi: https://doi.org/10.48550/arXiv.2306.16785.
}
\author{
Antoine Barbieri
}
