% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeQut-package.R
\docType{package}
\name{BeQut-package}
\alias{BeQut}
\alias{BeQut-package}
\title{BeQut: Bayesian Estimation for Quantile Regression Mixed Models}
\description{
Using a Bayesian estimation procedure, this package fits linear quantile regression models such as linear quantile models, linear quantile mixed models, quantile regression joint models for time-to-event and longitudinal data. The estimation procedure is based on the asymmetric Laplace distribution and the 'JAGS' software is used to get posterior sample (Yang, Luo, DeSantis (2019) \doi{10.1177/0962280218784757}).
}
\author{
\strong{Maintainer}: Antoine Barbieri \email{antoine.barbieri@u-bordeaux.fr}

Authors:
\itemize{
  \item Hélène Jacqmin-Gadda \email{helene.jacqmin-gadda@u-bordeaux.fr}
}

}
\keyword{internal}
